Switch back to form in flyout content, add tests
